Overhauled general object serialization
Taking advantage of new marshaller paradigm, object serializer does not have to concern itself with the types involved.
Also changed behaviour of deserialization so that a parameterless constructor for reference types is required otherwise an exception is thrown.
Also added an indicator for whether fields were serialized for a reference types (for value types, fields are always serialized).