Intermediate commit - extracted reference caching functionality
ReferenceTypeSerializer is now an abstract base class for all Serializers that want to serialize a reference type and preserve referential integrity. Pretty sure it's sufficient that each serializer has its own cache (as opposed to one shared between them), as long as serializer selection remains symmetrical on serialization/deserialization (which it should).