Changed contract naming
Names can now not be declared explicitly. That was a nice feature, but is highly impractical with generic types. Similarly fieldnames are now also inferred from the actual field name.
Also implemented a marshaller-level assembly/type resolver delegate, that can be used for version changes to somewhat replace the explicit contract naming.