Simplified serialization class hierarchy
Extracted the private serializationmarshaller class from serializer bass class, thereby making the imarshaller interface obsolete. The only realy change here is that Serializer.Marshaller's setter is now internal instead of private, but that's okay, it's pretty obvious that this backreference to marshaller shouldn't be fucked with, within the LEX assembly at least.